Move some files to be more consistent with main
