+++ 2024-Competition-Code/src/main/java/frc/robot/RobotContainer.java
@@ -8,10 +8,10 @@ import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.commands.*;
 import frc.robot.subsystems.PhotonVision.PhotonVision;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
-import frc.robot.commands.swerve.CrabDrive;
-import frc.robot.commands.swerve.SwerveAuto;
-import frc.robot.commands.swerve.SwerveTeleop;
-import frc.robot.commands.swerve.TestFourModules;
+import frc.robot.commands.CrabDrive;
+import frc.robot.commands.SwerveAuto;
+import frc.robot.commands.SwerveTeleop;
+import frc.robot.commands.TestFourModules;
 import frc.robot.subsystems.Shooter;
 import frc.robot.subsystems.swerve.SwerveDrive;
 import frc.robot.subsystems.swerve.SwerveModuleIO;
