+++ 2024-Competition-Code/src/main/java/frc/robot/commands/SwerveTeleop.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.swerve;
+package frc.robot.commands;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.geometry.Translation2d;
